Renamed core-lib to common-lib
