--- core-lib/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="ee.ria.DigiDoc.core" />
